tree_zig_zag - commits of README, HINTS, and SOLUTION
This commit contains the README description of the problem, a HINTS file
full of hints for when a user has already attempted the problem for some
time and requires some advice, and a SOLUTION human-readable summary of
the solution.

It is useful to break up this information into different files to avoid
giving spoilers too early to users who are attempting to simulate
interview-like scenarios.