tree_zig_zag - generator - initial commit
A script that can generate a random BST that is not balanced. This is
useful because it is helpful to have at least one large input to make
sure that a user's solution isn't too time inefficient.  Such a script
will probably come in handy for subsequent problems that require
unbalanced BSTs as test cases.